Ease setuptools requirements
Pretty stupid problem: for some uninvestigated reasons virtualenv
creates with setuptools==18.2 regarding that there is more actual
version installed on the site-packages.

This leads to the broken setup.py script as before setuptools==21
PEP-426 was not supported. But there is no graceful compatibility
degradation as well, so when some requirement with PEP-426 format
is met, error get throw and end of the story.

PKG mapping also changed to the official ports versions. The only
good side of this change as we don't have to build own package for
setuptools.

The mock test requirement is a bit redundant for Python 3.2+, but
I think this won't get hurt anyone for now. while setup options
will remain clean and simple.