Preserve full paths for archived files
Surprise, but tarfile strips leading slash without give a warning about.
This makes paths in MANIFEST and tar package inconsistent leading to
installation issues like "file is not included in manifest" and similar.

To overcome this problem, the only way it seems to get into `tar.add`
internals, get the tarinfo object and rename it with the original path.